requirements.txt
discord.py==2.3.2
rapidfuzz==3.9.0
python-dotenv==1.0.1
Flask==3.1.3
PyNaCl
aiohttp